fix: Create auth.ts
