+++ src/lib/eventAPI.ts
@@ -1,4 +1,5 @@
 import { api } from "@/lib/api";
+import { getAuthHeaders, getAuthToken } from "@/lib/auth";
 
 export type EventType = "firstdate" | "rpkm" | "freshmen-night" | "cufest";
 
@@ -25,27 +26,6 @@ export interface EventStatus {
   registrationSuccess?: boolean;
   isComingSoon?: boolean;
   checkinData?: CheckinResponse;
-}
-
-function getAuthToken(): string | null {
-  if (typeof window !== "undefined") {
-    const localToken = localStorage.getItem("auth_token");
-    if (localToken) {
-      return localToken;
-    }
-
-    const cookies = document.cookie.split("; ");
-    const tokenCookie = cookies.find((cookie) => cookie.startsWith("token="));
-    return tokenCookie ? decodeURIComponent(tokenCookie.split("=")[1]) : null;
-  }
-
-  return null;
-}
-
-// Helper function to create authorized headers
-function getAuthHeaders(token?: string): Record<string, string> {
-  const authToken = token || getAuthToken();
-  return authToken ? { Authorization: `Bearer ${authToken}` } : {};
 }
 
 export type ColorVariant =
